+++ api/src/lib.rs
@@ -36,7 +36,7 @@ use resources::{
 };
 use serde::{Deserialize, Serialize};
 use serde_json::json;
-use std::{cell::Cell, fmt::Display, path::Path};
+use std::{cell::Cell, fmt::Display, path::Path, time::Duration};
 use url::Url;
 
 use crate::resources::{
@@ -1844,10 +1844,14 @@ impl Endpoints {
     }
 }
 
+const DEFAULT_HTTP_TIMEOUT_SECONDS: u64 = 120;
+
 fn build_http_client(config: &Config) -> Result<HttpClient> {
     let mut builder = HttpClient::builder()
         .gzip(true)
-        .danger_accept_invalid_certs(config.accept_invalid_certificates);
+        .danger_accept_invalid_certs(config.accept_invalid_certificates)
+        .timeout(Some(Duration::from_secs(DEFAULT_HTTP_TIMEOUT_SECONDS)));
+
     if let Some(proxy) = config.proxy.clone() {
         builder = builder.proxy(Proxy::all(proxy).map_err(Error::BuildHttpClient)?);
     }